fix(#133): make certificate pinning web-compilable via conditional import
Address review feedback on certificate pinning:
- Split the pinning adapter into IO and web implementations selected by a
  conditional export (defaults to the web stub, picks dart:io only when
  available). This unblocks `flutter build web`, which could not compile the
  dart:io / dio/io.dart imports even though the runtime path was kIsWeb-guarded.
- Web build returns the default BrowserHttpClientAdapter (pinning is a no-op
  on web by design; browsers own TLS validation).
- Make the case-sensitivity pin test deterministic by selecting a fixture with
  a guaranteed uppercase base64 char instead of early-returning.

Verified with `flutter build web --target lib/main/main_prod.dart` (succeeds).
